Move github.com/labstack/echo example to godoc
