--- example/echo/go.mod
@@ -1,10 +1,0 @@
-module github.com/instana/go-sensor/example/echo
-
-go 1.15
-
-require (
-	github.com/instana/go-sensor v1.31.0
-	github.com/instana/go-sensor/instrumentation/instaecho v1.0.0
-	github.com/labstack/echo/v4 v4.6.0
-
-)
