--- example/echo/main.go
@@ -1,41 +1,0 @@
-// (c) Copyright IBM Corp. 2021
-// (c) Copyright Instana Inc. 2016
-
-// +build go1.15
-
-package main
-
-import (
-	"flag"
-	"os"
-
-	"github.com/instana/go-sensor/instrumentation/instaecho"
-
-	"github.com/labstack/echo/v4"
-
-	instana "github.com/instana/go-sensor"
-)
-
-var listenAddr string
-
-func main() {
-	flag.StringVar(&listenAddr, "l", os.Getenv("LISTEN_ADDR"), "Server listen address")
-	flag.Parse()
-
-	if listenAddr == "" {
-		flag.Usage()
-		os.Exit(2)
-	}
-	// create a sensor
-	sensor := instana.NewSensor("sensor")
-	// create an instrumented Echo instance
-	engine := instaecho.New(sensor)
-
-	engine.GET("/myendpoint", func(c echo.Context) error {
-		return c.JSON(200, map[string]string{
-			"message": "pong",
-		})
-	})
-
-	_ = engine.Start(listenAddr)
-}
